Move test to correct folder (?)
